+++ src/main/java/org/openlmis/fulfillment/i18n/MessageKeys.java
@@ -29,6 +29,7 @@ public abstract class MessageKeys {
   private static final String ORDER = "order";
   private static final String ORDER_RETRY = "orderRetry";
   private static final String ORDER_UPDATE = "orderUpdate";
+  private static final String ORDER_REQUISITION_LESS = "orderRequisitionLess";
   private static final String ORDER_CANCEL = "orderCancel";
   private static final String ORDER_FILE_TEMPLATE = "fileTemplate";
   private static final String REFERENCE_DATA = "referenceData";
@@ -105,6 +106,9 @@ public abstract class MessageKeys {
 
   public static final String ORDER_UPDATE_INVALID_STATUS =
       join(ERROR_PREFIX, ORDER_UPDATE, INVALID_STATUS);
+
+  public static final String ORDER_REQUISITION_LESS_ORDERABLE_NOT_FOUND =
+      join(ERROR_PREFIX, ORDER_REQUISITION_LESS, ORDERABLES, NOT_FOUND);
 
   public static final String ORDER_CANCEL_INVALID_STATUS =
       join(ERROR_PREFIX, ORDER_CANCEL, INVALID_STATUS);